requirements.txt
streamlit>=1.33,<2
openai>=1.0.0
aiohttp>=3.9.0
pydub>=0.25.1
audio_recorder_streamlit==0.0.6
SpeechRecognition>=3.10.0
PyPDF2>=3.0.1
python-docx>=0.8.11
pdf2image>=1.16.3
pytesseract>=0.3.10
Pillow>=10.2.0
numpy>=1.26.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
tiktoken>=0.7.0
